proyecto/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator 
python-multipart
pydantic-settings